fixed metrics tests
